Seperate debug action in debug.go
