+++ mvm/eos/contracts/dappdemo/build-debug.sh
@@ -1,0 +1,2 @@
+
+eosio-go build -tags debug -o dappdemo.wasm .
