+++ mvm/eos/contracts/mixincross/build-debug.sh
@@ -1,0 +1,2 @@
+
+eosio-go build -tags debug -o mixincross.wasm .
