+++ mvm/eos/contracts/mixincross/go.mod
@@ -2,4 +2,4 @@ module dappdemo
 
 go 1.17
 
-require github.com/uuosio/chain v0.1.4-0.20211227051229-9803aafbe6a3
+require github.com/uuosio/chain v0.1.5
