+++ mvm/eos/contracts/mtg.xin/build-debug.sh
@@ -1,0 +1,2 @@
+
+eosio-go build -tags debug -o mtg.xin.wasm .
