+++ mvm/eos/contracts/mtg.xin/go.mod
@@ -2,4 +2,4 @@ module mtg.xin
 
 go 1.17
 
-require github.com/uuosio/chain v0.1.2
+require github.com/uuosio/chain v0.1.7
